+++ lib/identity_cache/memoized_cache_proxy.rb
